requirements.txt
disnake==2.9.1
requests==2.31.0
vk==3.0
python-dotenv==1.0.0